started to initialize leaflet map in map component
